Modularize GitHub content service
Separate shared repository policy and content-file CRUD behind the existing route-facing facade without changing API contracts or behavior.
